Convert Spatial from float to double

+++ processing/src/main/java/org/apache/druid/collections/spatial/ImmutablePoint.java
@@ -49,7 +49,7 @@ public class ImmutablePoint extends ImmutableNode
     );
   }
 
-  public float[] getCoords()
+  public double[] getCoords()
   {
     return super.getMinCoordinates();
   }
